move aggregate stats to get_norm
Moved the aggregation functions for the stats that define normalization to get_norm_reducer.py. Also changed the keys for the output of first mapper. Instead of each key starting with 'features' or 'stats', these lines will be distinguished by whether their key is a channel name, or a filename-window#-channelname-label pair.